Run new unit test only against sqlite
The test_status_retrieve_sequence test, one of the new unit tests
introduced together with the partner sites feature from ALMA, runs
correctly only against the sqlite database, because fully separate
server instances are expected. This is a fundamental limitation of our
unit test infrastructure: when using MySQL and PostgreSQL only clustered
servers can be started.

To fix this situation we are skipping the problematic test when not
running against sqlite. An alternative solution in the future could be
to offer the option of forcing newly started servers to use sqlite,
therefore making them independent again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>