Move from Travis CI to GitHub Actions
Travis CI has become increasingly difficult to maintain and use,
specially after they moved to a business model where payment is the
default option, and open source projects like ours need to constantly be
contacting them to renew open source credit allocations (which also
don't always work as they should).

This commit moves the whole infrastructure for our CI tests from Travis
CI to GitHub Actions. Most of the steps performed during the tests were
already encapsulated in the run_{build,test}.sh scripts, but they still
needed small modifications to cater for the new environment they run in.

Two workflows are defined, and that mimic the two types of jobs we had
in Travis: one that makes sure the documentation builds correctly, and
another that builds NGAS and runs its unit tests in a number of
configurations. These configurations are similar to those previously
found in Travis, with 2.7 builds against different database backends,
and builds for the main 3.X releases (bar 3.10 and 3.11, which hadn't
been included in Travis either).

After this move, the collected coverage is extremely similar to that
collected from Travis, with the difference that the bbcp tests are
skipped, which takes the coverage percentage about half point down.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>