Fix --servers flag for ngamsArchiveClient
The ClientWrapper class in the fs_monitor module was always creating an
ngamsPClient object with host, port and servers, although they are
mutually exclusive arguments (host/port and servers). This was only
triggered when users specified the --servers flag on the
ngamsArchiveClient utility.

This commit fixes this situation, and adds a new unit test to ensure the
--servers flag continues working as expected.

This addresses #38.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>