Use aiosmtpd when smtpd is not available
The SMTP server used by tests has currently been based on the smtpd
system module. This module is deprecated though, as slated for removal
in python 3.10, as indicated in #14. If we want to continue testing our
mailing capabilities we therefore need to address this situation.

This commit contains the necessary changes to keep supporting our SMTP
test server using an external python package, aiosmtpd. The package is
the recommended way of moving away from smtpd, and even had a useful
guide on how to port code from one system to another. It also doesn't
require us to keep track of the background thread running the server,
which is a nice addition. Other than its API, it had one difference
though: messages containing newline characters (\n) appear on the other
side with a \r\n sequence, which didn't happen with smtpd. I haven't
invested time to investigate why this is different between the two
packages, but doing a small replace() call on the incoming data suffices
for our purposes.

The aiosmtpd package works only with python 3.6+, which is a narrow list
compared to the one we want to support in NGAS. Because of this we
therefore are maintaining both the smtpd- and the aiosmtpd-based
versions of the SMTP servers, and only when smtpd is not available
we switch to using the aiosmtpd-based solution.

Finally, aiosmtpd requires a coroutine to be implemented. This is of
course not possible in python 2.7 ("async def" doesn't parse). We avoid
this by declaring an awaitable in a more "raw" format (i.e., returning
an object with a __await__ method returning an iterator, etc...)

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>