Generate useful message upon unknown MIME type
As pointed out by #13, issuing an archiving command with an unknown MIME
type currently results on a cryptic, internal error message ("'NoneType'
object has no attribute 'getHostIdList'") rather than a useful one.

This commit avoids the condition that generated the previous error
message, adding the necessary checks to generate a more meaningful error
message instead stating that the MIME type is not konwn to NGAS.
Alongside this change I also slightly simplified a bit of code to avoid
some small repetitions.

A new unit test also ensures the error message received by the user is
actually useful. This is done by checking that the unknown MIME type
sent by the user is somewhere in the error message, hinting that the
issue is being addressed by the message.