Add HTTP status to ngamsStatus object
The ngamsStatus object so far has lacked information about the HTTP
status of the response that contained/generated it. This information is
useful in general, but in particular makes tests checking for the HTTP
status code easier to write.

This commit adds a new http_status member to the ngamsStatus that can be
initialised at __init__ time (defaults to None for backwards
compatibility), and adjusts some of the main ngamsStatus creation
routines to feed this information from their HTTP responses.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>