Remove MacOS builds from Travis
The MacOS builds in Travis have always been a bit of a problem, mostly
because installing packages and preparing the system is a bit of an
issue there. Moreover we don't really support MacOS as a deployment
target, it's just been a nice thing to test against for portability (and
some people do development in MacOS anyway).

This commit removes the MacOS job from the Travis matrix. The rest of
the MacOS-specific instructions in the Travis CI scripts is left
untouched though, in case someone wants to brings this up back to life.
In time that should be removed too though.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>