Fix logging in ngams C utilities
The core logging function in the ngams C library was broken: its output
contained only the timestamp, thread ID and log level, but not the
actual message and parameters. This was due to the wrong buffer being
used for outputting the latter information.

This patch fixes this situation, printing the user-provided parameters
into the buffer that is finally outputted to stdout and to the logfile.
In addition a newline character is appended at the end, as both sinks
need this additional character in order to become readable.

This addresses #41.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>