Finish implementation of -v flag in ngamsCClient
For some reason the -v flag of the ngamsCClient was not yet implemented,
even though it was just two small lines again from being done. The fact
that this wasn't implemented is, partly, why the error in the logging
functions of the ngams C library took this long to find out.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>