Inform crc variant on subscription push
When pushing data downstream, the subscription mechanism informed the
remote end about the file checksum value, but not the variant/type of
checksum used to calculate such value. This left the downstream server
incapable of working with checksums other than "crc32", while files
locally checksummed with other variants would fail to be pushed because
on the remote server the checksum values wouldn't match.

We vaguely knew this was a limitation from old discussions, but it
became a problem more explicitly in #44.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>