Improve error message sent back by REGISTER
MIME types might be configured in the server, but the list of REGISTER
command plug-ins can still be missing a plug-in for some of these MIME
types. The previous error message sent back by the server in these
situations was that of a KeyError, which is simply the missing key;
therefore all users saw was "<mime-type>" with the MIME type of the file
being registered.

A new unit test also checks that the error message sent back in these
situations contains a least the word "mime-type", which signals that
some information is being given to the user other than the MIME type
itself.

This addresses #47.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>