Remove test-specific generic register plug-in
This plug-in was actually very generic, and very similar (if not exactly
the same semantically) as the ngamsRegisterGenericPlugIn module provided
by Stewart in the previous commits. Now that the latter is in place we
don't need the former, so we can safely remove it. This also highlighted
that one of the new unit tests for the new generic plug-in was in fact a
copy of one of the original unit tests, so I removed the copy
referencing the now removed testing plug-in.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>