Allow commands to return (code, status, msg) tuples
We already allow command modules to return messages, which we interpret
as success messages. Sometimes commands want to send back a different
HTTP code and NGAS status with a message too, and allowing them to
return a simple 3-element tuple with this information is an easy way to
add this capability.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>