Improve handling of milliseconds in toiso8601
The way milliseconds were formatted in the toiso8601 function was
slightly problematic. Milliseconds correspond to the fractional part of
the number given as an input to the function. Because time formatting
functions don't support millisecond formatting, we do the formatting
ourselves by extracting the fractional part of the number, turning it
into an integer, and then printing it with three digits.

Because of how floating point numbers work, for certain inputs this
fractional part, when extracted, doesn't have exactly the same value,
and in particular one of the first three digits might change
(e.g., when extracting 1621405986.0 off 1621405986.063 the result is
0.06299996376037598). We were blind to this fact, and were always first
multiplying this fractional part by 1000 and then interpreting that as
an integer before performing the string formatting. This meant that for
these inputs we generated the incorrect number of milliseconds in the
generated string (".062" in the case above).

This commit fixes this problem by always rounding the fractional
remainder first, before multiplying by 1000 and interpreting that as an
integer. At the same time I've also switched to using math.modf for
obtaining the fractional part, which is a more trust-worthy method than
anything I can come up with.

This new method of getting millisecond formatting means we need to take
into account situations when the number of milliseconds round up to a
second (i.e., when msecs >= 0.9995). In such cases we adjust the values
to have an extra second with zero milliseconds printed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>