Notify clients when subscription exists
This change is the core of #50: it checks if a subscription for the
given ID exists or not before attempting insertion. If one exists, the
server replies to the client with different HTTP codes depending on
whether the subscription is identical to the one the client is trying to
add or not.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>