Handle existing subscriptions on server startup
Now that the SUBSCRIBE command lets us know if a subscription already
exists or not on the remote end, we can do things differently at server
startup to avoid continuous errors being thrown at us.

This commit adds the ability to check when a subscription already
exists. If it does, and its contents are different from the one we just
tried to create, an error is logged, an email is sent to the server
administrators, and no further attempts are done to register the faulty
subscription.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>