Address problems found during testing
Stewart tried out the latest changes to the subscription logic and found
three different problems, all of which are addressed in this commit.

Firstly, a typo in the email message has been fixed. Secondly, a logging
statement was missing a value for a format specifier. These two changes
were minor.

Finally, and most importantly, when a subscription fails to be added
because a similar-looking one already exists, the code tried to remove
the subscription object twice from the list of yet-to-be-created
subscriptions, which raised an unexpected error. However, at the same
time Stewart pointed out that at server shutdown the subscription that
wasn't added *was* removed from the remote server, which is what he
expected to see. This is not exactly great (if we are not adding it we
shouldn't be removing it), but on the other hand I see the appeal on
doing so, and since they are the main users of this feature anyway I'll
follow their preference. To avoid the double removal of the object,
while keeping the code path separate from the fully successful cases, I
created a small function that centralises the marking of a subscription
as active on the local server so it gets removed when the server goes
down.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>