Start communication thread before janitor process
The main server process and the janitor process communicate through a
pair of Queues. The handling of this communication on the main server
process happens on a separate thread, which we used to start after we
spawned the janitor process. This sequence opened the door for a race
condition where the janitor process would try to read from the
corresponding queue, but the main server wouldn't have put any item on
it yet as the communication thread wouldn't have started yet.

This commit fixes this situation by starting the communication thread
before spawning the janitor process. Since we are not waiting for the
communication thread to be up and running (only starting it) there is
still some chance that the original problem could occur, but with a much
lesser probability.

This addresses the problem originally reported in #78.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>