Update to Ubuntu focal as base platform
This update brings two new issues that need to be solved. First, the
MySQL-python package needs to be manually given a my_config.h file to
compile, as the base distro doesn't support the package anymore. And
secondly, the setup to get PostgreSQL running is slightly different,
needing to install the apt packages, starts the postgresql addon
(instead of the service), and needing to point to the database in
non-standard ports and with the travis user.

All in all it's not a big change, and we are now standing on a newer
distro for future builds, saving us from stagnating a bit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>