Ensure returned values are always float
The secs variable can either be a float or an integer, and depending on
this the final result of this function can be of either type too. The
documentation says this function always returns a float though, and it's
good to stay consistent, so this commit ensures the result is always of
type float.

This problem was originally reported in #87.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>