Use extra variable for data of different type
Static analysis tools complain if the same variable name is used to
point to data of a different types. In this case a list of strings was
coalesced into a single string, but the same variable name was
retained, triggering complains from mypy.

This commit introduces a new variable to keep track of these pieces of
data separately. The penalty of keeping a reference to the original list
of strings around is minimal.

This was reported in #86.