Initialise members with consistent type
The __ioTime member of ngamsDapiStatus is always used as a float value,
but it is initialised as an integer. While not a problem on itself, it
can be confusing; moreover static analysis tools complain about it.

This commit initialises this member to a float value, removing any
confusion.

This was reported in #85.