Use base64.b64{encode,decode} for python 3 support
The old {encode,decode}string functions were marked as deprecated in
python 3.1, and removed in python 3.9. We should use the corresponding
b64{encode,decode} functions instead, which are present in all relevant
python versions we support.

Note that in python 3 the b64{encode,decode} functions require bytes as
inputs, and output bytes (in python 3); we thus must take care of
encoding the arguments and decoding the outputs, while maintaining
python 2 compatibility.

Another small difference is that encodestring() returned a string that
always finished with a newline, which decodestring() handled correctly;
this is not the case with the b64{encode,decode} pair: b64encode()
doesn't append a newline, and b64decode() fails when a newline is
present at the end of its input. This won't have a noticeable effect
though: the encrypt_access_code function that issued the call to
encodestring() is never invoked anyway, and the strings given to
decrypt_access_code is always stripped (i.e., it won't contain
newlines).

This was originally reported in #89.