Check presence of all DB parameters in RC file
Some of the parameters the NGAS utilities read from the RT file can be
missing, as these files are written by users. No error is normally
raised When such parameters are not found (None is returned instead),
which can lead to difficult-to-spot errors down the line.

This commit adds a new flag to the get_parameter_ngas_resource_file
function to raise an error when a required parameter is not present
rather than return None. It defaults to False (to preserve backwards
compatibility), and is set to True in all invocations performed from the
get_db_parameters function, which is the function that originated this
improvement.

This was requested in #92.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>