Open file for reading in bytes mode
The hashing functions work with bytes, not strings, and therefore we
must supply them with the correct data type. This wasn't an issue in
python 2, but since the default opening mode for files is text (yielding
strings) it caused issues with python 3.

This addresses one of the issues raised in #93.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>