Fix problem with logfile removal logic
The logic for old logfile removal was flawed, as it was removing the
*newest* rotated logfiles instead of the newer ones. This was spotted by
Stewart, who observed old logfiles not being removed when they should,
and noted that logfile rotation was apparently not working either
(although this last point seems to be a confusion stemming from the fact
that newer rotations were deleted straight away).

The problem was a wrong slicing operation on the rotated logfiles list.
This is being fixed in this commit, along with an improvement on the
information being logged when performing this operation to make it
clearer to people reading the logs in the future what exactly is going
on.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>