Allow repeated starts of background tasks
Tasks are given an Event object to be used as a timely stopping
mechanism. At construction time that Event is clear, and is set at
stop() time for the background task to react to this. However, we never
cleared it back, meaning that calling start() on a Task started the
background logic, which would most probably quickly return due to the
event being already set.

This commit clears the Event object before starting the background task
every time, ensuring its logic runs to completion. A new unit test
checks that this is working correctly.

This addresses the problem described in #97, where the Janitor process
was defunct after an OFFLINE + OFFLINE series of commands. This happened
because the Task object that managed the janitor process didn't clear
the Event before starting the background task again, which in turn made
the new Janitor process exit quickly upon seeing the Event set. The
process was in defunct state because the server hadn't tried to join it
yet, which would have happened at OFFLINE/shutdown time anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>