Add support for "ignore" column to QUERY command
The QUERY command created its supported queries at module loading time,
at which time it didn't know if the server was configured to use the
"file_ignore" column name or the older "ignore" column name for the
column in the "ngas_files" table; instead it assumed "file_ignore"
throughout.

This commit adds support to the QUERY command to be able to use the
older "ignore" column name. To do this we now construct the list of
supported queries at runtime, when the first query comes in, and when we
have a hold on the DB object that knows which column to use.

To ease this change I've added a property to the DB object to easily
query the name of this column from outside the class.

This commit addresses #98.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>