Return ngamsStatus.dtd when requested
In 3434432 we commented out all "internal" RETRIEVE requests, since the
mechanism used to fulfill them was too generic and opened the door for
obtaining sensitive information to whomever had access to send requests
to the server. Then in cb42a47 we nailed the coffin and removed that
commented code. However, it was pointed out to us that this change
removed the ability to perform XML validation of returned XML documents,
since the ngamsStatus.dtd schema was made available through this
mechanism.

This commit brings back the support for the "internal" HTTP parameter,
but *only* when its value is "ngamsStatus.dtd", which returns the DTD
document. Any other value results in the "parameter not supported
anymore" message that we introduced a while back.

This problem was reported in #103.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>