Improve generation of DOCTYPE declarations
DOCTYPE declarations pointed to the RETRIEVE command of the server with
the correct flags that would trigger the retrieval of the top-level DTD
file that can be used to validate the ngamsStatus XML documents. The URL
pointed to by the DOCTYPE declaration though was not using the Host HTTP
header sent by the client, which is the best way to generate references
to other resources in the same server with a URL that the client can
use; instead, they were being generated using the fully-qualified
hostname of the machine hosting the server, which doesn't always map to
the hostname used by the client to contact the server (e.g., when the
server is behind a NAT).

The addStatusDocTypeXmlDoc call has now been simplified to take the XML
string and the host to inject into the DOCTYPE URL; otherwise it's the
responsibility of the caller to provide a useful name, which in most
cases is readily available through the HTTP handler's host property.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>