Enrich archive_event with more information
We previously only passed a file's ID and version to the plug-ins
through this event. We now pass the disk ID and the server itself,
allowing plug-ins to perform much more complex operations when they
receive these events.

This required a small change in one of the test handlers that pickled
the event, which isn't possible anymore.

This small development was done as part of NGAS-164.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>